backend/requirements.txt
requests
python-dotenv
geopandas
shapely
